Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int n;
+    float n;
     scanf("%f",&n);
     printf("Area: %.2f",(3.14)*n*n);
     return 0;
